median: Make TupleSort context hang from non-volatile memory context
The memory context allocated within tuplesort_begin_common hangs from
the current memory context at that moment. During the allocation, the
parent of "TupleSort main" context is set to ExprContext, which is a
sort of volatile context which gets reset for every tuple. From the doc:

  The upshot of that reasoning, though, is that CurrentMemoryContext
  should generally point at a short-lifespan context if at all possible.
  During query execution it usually points to a context that gets reset
  after each tuple.

Subsequent calls to tuplesort_putdatum see no parent for
"TupleSort main", which causes trouble when resizing the collection of
tuples (grow_memtuples in tuplesort.c).

In order to avoid this, the current memory context which will be set
as the parent of "TupleSort main", needs to be switched to the parent
of ExprContext prior to the call to tuplesort_begin_datum.